added Mockscreens for everything but "Multiplayer" Screens have a working "Back" button reorganized folder structure for better visibility
