+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,10 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".MainMenu" />
+        <activity android:name=".screens.MainMenu" />
+        <activity android:name=".screens.Settings" />
+        <activity android:name=".screens.Highscore" />
+        <activity android:name=".screens.Singleplayer" />
     </application>
 
 </manifest>
